+++ src/FSharp.Editing/Common/Lexer.fs
@@ -48,7 +48,7 @@ module Lexer =
                 | Some _, newLexState ->
                     loop lineTokenizer newLexState
 
-            let sourceTokenizer = SourceTokenizer(defines, "/tmp.fsx")
+            let sourceTokenizer = SourceTokenizer(defines, Some "/tmp.fsx")
             let lines = String.getLines source
             let lexState = ref 0L
             for line in lines do 
@@ -81,7 +81,7 @@ module Lexer =
         let defines =
             args |> Seq.choose (fun s -> if s.StartsWith "--define:" then Some s.[9..] else None)
                  |> Seq.toList
-        let sourceTokenizer = SourceTokenizer(defines, "/tmp.fsx")
+        let sourceTokenizer = SourceTokenizer(defines, Some "/tmp.fsx")
         let lineTokenizer = sourceTokenizer.CreateLineTokenizer lineStr
         let rec loop lexState acc =
             match lineTokenizer.ScanToken lexState with